+++ js/currentState.js
@@ -1,6 +1,8 @@
 export default {
-  answerCount: 0,
+  answerCount: 10,
   rightAnswerCount: 0,
+  livesLeft: 3,
   time: 1000,
+  timeLeft: 111,
   statistic: 77
 };
