+++ js/gameRules.js
@@ -1,5 +1,6 @@
 export default {
   gamesNumber: 10,
+  lives: 3,
   time: 2,
   artistsNumber: 3,
   genreNumber: 4,
