+++ js/main.js
@@ -2,6 +2,6 @@ import './animate.js';
 import './player.js';
 import './time-format.js';
 import './timer.js';
-import renderWelcome from './renderWelcome';
+import renderWelcome from './render/renderWelcome';
 
 renderWelcome();
