+++ js/render/renderWin.js
@@ -1,8 +1,8 @@
-import render from './render';
+import render from '../functions/render';
 import renderWelcome from './renderWelcome';
-import currentState from './currentState';
-import screenResult from './screenResult';
-import screenWin from './screenWin';
+import currentState from '../currentState';
+import screenResult from '../templates/screenResult';
+import screenWin from '../templates/screenWin';
 
 export default () => {
   render(screenResult(screenWin(currentState)));
