+++ js/templates/screenLose.js
